crosswalk: Ensure native programs run within the right environment.
Several tools are expected to be available on the build host to build
Crosswalk. Apart from those already required by Yocto, they should be
built on the fly, and be staged in the sysroot for build host. Moreover,
environment variables should be configured properly to ensure the right
search paths are used by these tools. Inheriting certain classes helps
to achieve all the above in an easy and reliable way.