crosswalk: Add missing depends.
Add missing dependencies based on the packaging scripts of popular Linux
distributions:

1) https://git.archlinux.org/svntogit/packages.git/tree/trunk/PKGBUILD?h=packages/chromium&id=6937f0369bb52bea6784c3ba7a9c0fa6f3e44aa4
2) http://anonscm.debian.org/cgit/pkg-chromium/pkg-chromium.git/tree/debian/control?h=debian/49.0.2623.108-2&id=7ec2b22090029969accc5f5b0cd2b993e442389b
3) http://pkgs.fedoraproject.org/cgit/rpms/chromium.git/tree/chromium.spec?id=098c7ea5bda01d99a9ac92b1b2e135f7c0929669
4) https://gitweb.gentoo.org/repo/gentoo.git/tree/www-client/chromium/chromium-49.0.2623.108.ebuild?id=fac9a5359578eba632d5112207e5906fa59426c0
5) https://build.opensuse.org/package/view_file/openSUSE:Factory/chromium/chromium.spec?expand=1&rev=d68a8add5bb89afc52b926b7b9ca7308
6) http://bazaar.launchpad.net/~chromium-team/chromium-browser/trusty-working/view/1113/debian/control
7) https://github.com/OSSystems/meta-browser/blob/77736988073a5d90fcff9d0005c8477332ede387/recipes-browser/chromium/chromium.inc

The following rules are followed to decide whether a dependency should
be added to the recipe:

1) Only those listed as build-time dependencies are considered, and
runtime dependencies are ignored. This is based on the assumption that
all runtime dependencies have already been added to DEPENDS and RDEPENDS
before by examining certain files.
2) There's no easy way to tell if a package that is mainly composed of
shared libraries or files for development is used on the build host or
it's just a runtime dependency on the target, so such kind of packages
are always ignored.
3) Script module dependencies are ignored because they are bundled with
the source and there's no way to use the system copy without additional
patches.
4) For other dependencies, if it is not treated as a build-time
dependency by most distributions, it is ignored.