crosswalk: Make pkg-config work as expected.
It's not necessary to explicitly tell pkg-config where the *.pc files
for target system are placed. gyp_xwalk can find them based on the
sysroot path. On the contrary, gyp_xwalk expects pkg-config to work as a
native tool. Therefore, environment variables exported by BitBake must
be unset to restore the default behavior of pkg-config. The only
exception is the search path of pkg-config. The PKG_CONFIG_LIBDIR
environment variable forces pkg-config to search for *.pc files in the
staging sysroot directory for build host, so that the result will not
get polluted by those shipped with build host.